Finalizado a tela do carrinho.

+++ S15_ErudioMicroservices.NET6-ImplementandoTeladeCarrinhodeCompras/GeekShopping/GeekShopping.Web/Models/CartHeaderViewModel.cs
@@ -6,5 +6,5 @@ public class CartHeaderViewModel
     public string UserId { get; set; }
     public string CouponCode { get; set; }
 
-    public double PurchaseAmount { get; set; }
+    public decimal PurchaseAmount { get; set; }
 }
